+++ lib/app/waydir_shell.dart
@@ -25,13 +25,16 @@ import '../core/models/file_operation.dart';
 import '../core/settings/settings_store.dart';
 import '../core/terminal/shell_detector.dart';
 import '../core/terminal/terminal_launch.dart';
-import '../core/update/update_store.dart';
 import '../features/containers/container_store.dart';
 import '../features/containers/wsl_path.dart';
 import '../features/checksum/checksum_dialog.dart';
+import '../features/command_palette/app_command.dart';
+import '../features/command_palette/command_palette_launcher.dart';
+import '../features/command_palette/command_palette_view.dart';
+import '../features/command_palette/command_usage_store.dart';
 import '../features/compare/compare_mode_bar.dart';
+import '../features/drives/drive_store.dart';
 import '../features/help/help_dialog.dart';
-import '../features/update/update_dialog.dart';
 import '../features/navigation/bookmark_store.dart';
 import '../features/navigation/navigation_store.dart';
 import '../features/navigation/sidebar.dart';
@@ -48,6 +51,7 @@ import '../features/plugins/plugin_icons.dart';
 import '../features/plugins/plugin_models.dart';
 import '../features/plugins/plugin_settings_store.dart';
 import '../features/plugins/plugin_store.dart';
+import '../features/settings/keybinding_labels.dart';
 import '../features/settings/preferences_view.dart';
 import '../features/tags/tag_edit_dialog.dart';
 import '../features/tags/tag_path.dart';
@@ -73,6 +77,7 @@ part 'waydir_shell/base.dart';
 part 'waydir_shell/actions.dart';
 part 'waydir_shell/terminal.dart';
 part 'waydir_shell/menus.dart';
+part 'waydir_shell/command_palette.dart';
 part 'waydir_shell/keyboard.dart';
 
 class WaydirShell extends StatefulWidget {
@@ -88,6 +93,7 @@ class _WaydirShellState extends State<WaydirShell>
         _WaydirActionsMixin,
         _WaydirTerminalMixin,
         _WaydirMenuMixin,
+        _WaydirCommandPaletteMixin,
         _WaydirKeyboardMixin {
   @override
   void initState() {
@@ -186,9 +192,9 @@ class _WaydirShellState extends State<WaydirShell>
         _installRenameErrorEffects();
       }),
     );
-    _installUpdateNotification();
     _installPluginEventEffects();
     _maybePromptFullDiskAccess();
+    CommandPaletteLauncher.instance.open = _openCommandPalette;
   }
 
   void _maybePromptFullDiskAccess() {
@@ -203,6 +209,9 @@ class _WaydirShellState extends State<WaydirShell>
 
   @override
   void dispose() {
+    if (CommandPaletteLauncher.instance.open == _openCommandPalette) {
+      CommandPaletteLauncher.instance.open = null;
+    }
     for (final d in _effectDisposers) {
       d();
     }