+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 987
+/// Strings: 994
 ///
-/// Built on 2026-06-21 at 21:34 UTC
+/// Built on 2026-06-22 at 15:28 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import